refactor: `TxLayout`. Merge all tx page files
