--- source/pages/Transactions/CreateNFT.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import { TxConfirmLayout, TxLayout } from 'components/Layout';
-
-// confirmCreateNFT
-export const CreateAndIssueNFTConfirm = () => (
-  <TxConfirmLayout title="NFT CREATION" txType="newNFT" />
-);
-
-export const CreateAndIssueNFT = () => (
-  <TxLayout
-    confirmRoute="/external/tx/asset/nft/issue/confirm"
-    txType="newNFT"
-    title="Create NFT"
-  />
-);
