--- source/pages/Transactions/SignPSBT.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import { TxConfirmLayout } from 'components/Layout';
-
-export const SignPSBT = () => (
-  <TxConfirmLayout sign title="SIGNATURE REQUEST" txType="signPSBT" />
-);
