--- source/pages/Transactions/TransferOwnership.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import { TxConfirmLayout, TxLayout } from 'components/Layout';
-
-// confirmAssetTransfer
-export const TransferOwnershipConfirm = () => (
-  <TxConfirmLayout title="TRANSFER ASSET" txType="transferAsset" />
-);
-
-export const TransferOwnership = () => (
-  <TxLayout
-    confirmRoute="/external/tx/asset/transfer/confirm"
-    txType="transferAsset"
-    title="Transfer Asset"
-  />
-);
